Move all the branch checking into its own file
